requirements.txt
discord.py==1.7.3
Flask==2.2.5
requests==2.31.0
openai==0.27.0
python-dotenv==1.0.0